% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_gifs.R
\name{combine_gifs}
\alias{combine_gifs}
\title{Combines two gifs side by side}
\usage{
combine_gifs(a, b, outfile = NA)
}
\arguments{
\item{a}{gif a, either an output directly from gganimate or the path to a gif (faster)}

\item{b}{gif b, either an output directly from gganimate or the path to a gif (faster)}

\item{outfile}{if the file should directly be written to a gif}
}
\value{
an magick-image
}
\description{
Combines two gifs side by side
}
\examples{
sh <- run_schelling()
\donttest{
  grid <- plot_grid(sh, animate = TRUE, title = TRUE)
  dev  <- plot_development(sh, animate = TRUE)
  combined <- combine_gifs(grid, dev)
  print(combined)
}
}
