% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_development.R
\name{plot_development}
\alias{plot_development}
\title{Plots the development of happiness over time}
\usage{
plot_development(d, animate = FALSE, step = 1, title = TRUE)
}
\arguments{
\item{d}{a list of data-frames as returned by run_schelling}

\item{animate}{if gganimate is installed, if the plot should be animated}

\item{step}{if animate = TRUE, which steps should be plotted, either a vector or a single number}

\item{title}{if the title should be displayed}
}
\value{
a ggplot2 object, or a gganimate object if animate = TRUE
}
\description{
Plots the development of happiness over time
}
\examples{
sh <- run_schelling()
plot_development(sh)

\donttest{
 # if the gganimate package is installed:
 plot_development(sh, animate = TRUE)
}
}
