% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grid.R
\name{plot_grid}
\alias{plot_grid}
\title{Plots the grid of a simulation}
\usage{
plot_grid(d, select_round = 0, grid = TRUE, title = FALSE,
  emoji = FALSE, emoji_size = 7, animate = FALSE, step = 1)
}
\arguments{
\item{d}{a list of data-frames as returned by run_schelling}

\item{select_round}{the round number to display}

\item{grid}{bool value if the grid should be displayed}

\item{title}{if the title should be displayed}

\item{emoji}{if emojifont is installed, if the state of the agents should be displayed using emojis}

\item{emoji_size}{size of the text of the emojis (only used if emoji = TRUE)}

\item{animate}{if gganimate is installed, if the plot should be animated}

\item{step}{if animate = TRUE, which steps should be plotted, either a vector or a single number}
}
\value{
a ggplot2 object, or a gganimate object if animate = TRUE
}
\description{
Plots the grid of a simulation
}
\examples{
sh <- run_schelling()
plot_grid(sh)
plot_grid(sh, Inf, title = TRUE)

\donttest{
 # if the emojifont package is installed:
 plot_grid(sh, emoji = TRUE)

 # if the gganimate package is installed:
 plot_grid(sh, animate = TRUE)
 plot_grid(sh, animate = TRUE, step = 2, title = TRUE)

 # if both are installed:
 plot_grid(sh, emoji = TRUE, animate = TRUE, title = TRUE)
}
}
