% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_schelling.R
\name{run_schelling}
\alias{run_schelling}
\title{Runs the Schelling ABM for given sets of parameters}
\usage{
run_schelling(size = 10, percent_empty = 0.2, threshold = 0.5,
  number_of_groups = 2, max_rounds = 1000, seed = runif(1, 0L,
  .Machine$integer.max))
}
\arguments{
\item{size}{the size of the grid (i.e., an size x size square)}

\item{percent_empty}{how many of the spaces should be empty?}

\item{threshold}{under which threshold should agent become unhappy and move}

\item{number_of_groups}{how many different groups should be simulated}

\item{max_rounds}{what is the maximum number of simulational rounds allowed}

\item{seed}{a seed for the random number generator (Mersenne-Twister)}
}
\value{
a list of two data_frames containing the simulational results
}
\description{
Runs the Schelling ABM for given sets of parameters
}
\examples{
library(dplyr)
library(ggplot2)

sh <- run_schelling()
sh

sh$round \%>\%
  mutate(perc_happy = happy_agents / (happy_agents + unhappy_agents)) \%>\%
  ggplot(aes(x = round, y = perc_happy)) + geom_line()

sh$detailed \%>\%
  filter(group != 0) \%>\%
  filter(round == 1) \%>\%
  ggplot(aes(x = x, y = y, fill = as.factor(group))) +
  geom_tile() +
  geom_point(aes(shape = as.factor(happy)), size = 6)

sh$detailed \%>\%
  # group 0 stands for empty
  filter(group != 0) \%>\%
  group_by(round, group) \%>\%
  summarise(perc_happy = mean(happy)) \%>\%
  ggplot(aes(x = round, y = perc_happy, color = as.factor(group))) +
  geom_line()

\donttest{
# takes a while, but has a nice pattern :D
sh <- run_schelling(100, percent_empty = 0.2, threshold = 0.5,
                    number_of_groups = 10, max_rounds = 20000)

sh$round \%>\%
  mutate(perc_happy = happy_agents / (happy_agents + unhappy_agents)) \%>\%
  {plot(.$round, .$perc_happy, type = "l")}

sh$detailed \%>\%
  filter(group != 0) \%>\%
  group_by(round, group) \%>\%
  summarise(perc_happy = mean(happy)) \%>\%
  ggplot(aes(x = round, y = perc_happy, color = as.factor(group))) +
  geom_line()
}
}
